+++ src/services/encryption/symmetric-encryption.ts
